Send text to Arduino through BT with python
Requires the pybluez library, which is easy to install on Ubuntu
but much more difficult on windows. The best bet for 64bit python
is through precompiled wheels:
http://www.lfd.uci.edu/~gohlke/pythonlibs/#pybluez

The UUID should be the same for all HC-05 modules since it's a
function of its serial port class property. The address is specific
to this module, but the address of any can be found with
the print_nearby_devices function.